🚚 rename: Reservation page 폴더위치 변경 #91
